### 2022-10-27 #559 @schneiderpy
concstats
Submitting Author Name: Name
Submitting Author Github Handle: @schneiderpy
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/schneiderpy/concstats
Version submitted:
Submission type: Stats
Badge grade: silver
Version accepted: TBD
---

```
Package: concstats
Title: Market Structure, Concentration and Inequality Measures
Version: 0.1.4
    person(given = "Andreas",
           family = "Schneider",
           email = "schneiderconsultingpy@gmail.com",
           comment = c(ORCID = "https://orcid.org/0000-0001-5630-1097"))
Description: Based on individual market shares of all participants in a
    market or space, the package offers a set of different structural and
    concentration measures frequently - and not so frequently - used in research
    and in practice. Measures can be calculated in groups or individually.
    The calculated measure or the resulting vector in table format should help
    practitioners make more informed decisions. Methods used in this package are
    from:
    1.  Chang, E. J., Guerra, S. M., de Souza Peñaloza, R. A. & Tabak, B. M.
    (2005) "Banking concentration: the Brazilian case".
    2.  Cobham, A. and A. Summer (2013). "Is It All About the Tails? The
    Palma Measure of Income Inequality".
    3.  García Alba Iduñate, P. (1994). "Un Indice de dominancia para el
    analisis de la estructura de los mercados".
    4.  Ginevicius, R. and S. Cirba (2009). "Additive measurement of market
    concentration" <doi:10.3846/1611-1699.2009.10.191-198>.
    5.  Herfindahl, O. C. (1950), "Concentration in the steel industry
    (PhD thesis)".
    6.  Hirschmann, A. O. (1945), "National power and structure of foreign trade".
    7.  Melnik, A., O. Shy, and R. Stenbacka (2008), "Assessing market dominance"
    <doi:10.1016/j.jebo.2008.03.010>.
    8.  Palma, J. G. (2006). "Globalizing Inequality: 'Centrifugal' and
    'Centripetal' Forces at Work".
    9.  Shannon, C. E. (1948). "A Mathematical Theory of Communication".
    https//schneiderpy.github.io/constats (website)
    readr
    devtools,
    kableExtra,
    covr
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in #558

The goal of the concstats package is to offer a set of alternative and/or additional measures for researchers in social sciences and practitioners in institutions concerned with competition on a regular basis to better determine a given market structure and therefore reduce uncertainty with respect to a given market situation.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

Some functions are already implemented in other R packages. The non-exhaustive
summary below is by no means a description of each package.

- The Herfindahl Hirschman Index can be found in the [hhi](https://joss.theoj.org/papers/10.21105/joss.00828) and the
[divseg](https://github.com/christopherkenny/divseg) packages. While the *hhi* package has just one function, neither of both packages offer a finite sample correction.

- The latter offers as well functions for the Entropy, Gini and Simpson measures.

- The [acid](https://cran.r-project.org/package=acid) and the [ineq](https://cran.r-project.org/package=ineq) packages offer functions for inequality and competition measures, e.g. for the Entropy and Gini metric.

However, almost none of these packages offer finite sample correction, with the exception of the
 [ineq](https://cran.r-project.org/package=ineq) package.
Other functions are new implementations in R, e.g. Dominance Index, Palma ratio, Stenbacka Index, GRS measure, and the dual of the Herfindahl Hirschman Index.

silver
* specific standards > 47%, and general standards > 58%
* fulfilling with two or more standards in the sub-categories 6.1.1, 6.1.2, and 6.1.5
