### 2022-10-30 #559 @mpadge
@schneiderpy Our usual procedure is to first consider suitability in a pre-submission enquiry. You've opened #558 for that, so we'll close this until the conversation is resolved there. Once that is done, and if the pre-submission is approved, we can re-open this issue and move on from there. Thanks.
