### 2022-10-31 #559 @adamhsparks
Thanks, @schneiderpy, as @mpadge noted, usually we like to finalise the pre-submission inquiry before opening a request for a review, but everything seems to be in top shape, so I've reopened this and will start looking for a handling editor.
