### 2022-11-22 #559 @annakrystalli
Hello @schneiderpy. It's because it's still at review stage 0. It will appear in the newsletter when an editor is assigned and it reaches stage 1/editorial checks 😊
