### 2023-01-20 #559 @schneiderpy
Hello Anna,

this is just a short email to find out if there is a chance the my package
"concstats" will get reviewed (someday)?

Best regards and an excellent 2023
Andreas Schneider

[image: Mailtrack]
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality11&>
Remitente
notificado con
Mailtrack
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality11&>
20/01/23,
12:27:20

El mar, 22 nov 2022 a las 10:25, Anna Krystalli ***@***.***>)
escribió:

> Hello @schneiderpy <https://github.com/schneiderpy>. It's because it's
> still at review stage 0. It will appear in the newsletter when an editor is
> assigned and it reaches stage 1/editorial checks 😊
>
