### 2023-02-01 #559 @schneiderpy
Thank you Mauro

[image: Mailtrack]
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality11&>
Remitente
notificado con
Mailtrack
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality11&>
01/02/23,
17:59:41

El mié, 1 feb 2023 a las 11:09, Mauro Lepore ***@***.***>)
escribió:

> Dear @schneiderpy <https://github.com/schneiderpy>,
>
> Today starts my rotation as EiC -- meaning I now assume the role of
> @annakrystalli <https://github.com/annakrystalli>.
>
> I'm watched the conversation here and among editors and it seems that the
> search for a handling editor continues. Thanks for your patience.
>
