### 2023-02-02 #559 @schneiderpy
Dear Mauro
Thank you for the good news. And of course, thank you to @Paula-Moraga
and @mark
padgham ***@***.***>

Best regards

[image: Mailtrack]
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality11&>
Remitente
notificado con
Mailtrack
<https://mailtrack.io?utm_source=gmail&utm_medium=signature&utm_campaign=signaturevirality11&>
02/02/23,
17:57:48

El jue, 2 feb 2023 a las 6:55, Mauro Lepore ***@***.***>)
escribió:

> Dear @schneiderpy <https://github.com/schneiderpy>,
>
> I'm pleased to announce that @Paula-Moraga
> <https://github.com/Paula-Moraga> will be the handling editor.
>
> Thanks @schneiderpy <https://github.com/schneiderpy> for your patience,
> @Paula-Moraga <https://github.com/Paula-Moraga> for accepting the role,
> and @mpadge <https://github.com/mpadge> for connecting all of us.
>
