### 2023-02-03 bbcab8f
Update README.Rmd
added rOpenSci badge for review
changed CRAN checks badge due to error
