+++ pkgdown.yml
@@ -1,8 +1,8 @@
-pandoc: 2.7.3
-pkgdown: 2.0.6
+pandoc: 2.19.2
+pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   concstats-intro: concstats-intro.html
   concstats-start: concstats-start.html
-last_built: 2022-10-26T14:00Z
+last_built: 2023-02-03T21:25Z
 
