### 2023-02-07 #559 @SebastianWojcik86
Dear all. It is my first review. I wanted to do it  in an exhaustive way and hope it shall be valuable for the Author.

---

#### General Standards for Statistical Software

G1.0 Statistical Software should list at least one primary reference from published academic literature.

In the list of references in the DESCRIPTION file I have not found Stenbacka index and Simpson. Please indicate where I can find it.

G1.1 Statistical Software should document whether the algorithm(s) it implements are...

Not stated in README file. I guess it should be "The first implementation within R of an algorithm which has previously been implemented in other languages or contexts". I have seen a similar package in python.

All other standards are met or irrelevant for the package.

#### Exploratory Data Analysis and Summary Statistics

All standards are met or irrelevant for the package.

#### Methodology

An argument *unbiased* explanation as 'Logical. Argument specifying whether or not a finite sample correction should be applied. The default is FALSE.' is misleading. Some measures can be normalized since their value - under equal shares assumption - depends on the number of market participants. Please correct the description of argument *unbiased* towards idea of normalization of the measure.

There is an inconsistency between required form of 'x' vector. For entropy, shares are calculated from 'x' while for HH index, input is in a form of shares. It should be unified.

Correct me, if I am wrong. Simpson index is defined in the package as a complement of HH index, that is $S=1-HH$. Under equal shares assumption it gives $S=1-\frac{1}{n}$. Following normalization of HH index, normalized Simpson index should be of the form
$$S_{norm}=\frac{S-1+\frac{1}{n}}{1-\frac{1}{n}}=\frac{\frac{1}{n}-HH}{1-\frac{1}{n}}.$$
It is not in a line with definition of normalized Simpson index in the package. Check simple calculations
```{r message=FALSE, warning=FALSE}
library(concstats)
# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# raw Simpson index
simpson(share)
# normalized Simpson index from the package
simpson(share,unbiased = T)
# normalized Simpson index? Make it sense?
(simpson(share)-(1-1/length(share)))/(1-1/length(share))
```

Function `grs` is implemented in a slightly different way then the way indicated in the reference. GRS does not to be restricted to positive shares. And so, `sum(x > 0)` should be replaced by `length(x)` or something similar. Such replacement has an impact on the result.

Function `dom` has no normalization option. Correctly?

#### Man files

- no citations included
- for Stenbacka index, I have not found a relevant reference. Is it defined just for two biggest market shares?. I guess that one is fine: Melnik, A., Shy, Oz, Stenbacka, R., (2008), “Assessing market dominance”, Journal of Economic Behavior and Organization 68, pp. 63-72.
- the most of measures can be described briefly with single formula and understood by majority package users. I encourage the Author to enrich help pages with such description using TEX e.g. *eqn* and *deqn* environments. It looks nice :)  $$HHI(x)=\sum_{i=1}^{n}\left(\frac{x_{i}}{\sum_{i=1}^{n}x_{i}}\right)^2$$
- description 'computes the numbers equivalent' of the function `nrs_eq` is very unclear
- in my opinion examples should be written to print outputs and informative, that is
```{r message=FALSE, warning=FALSE}
library(concstats)
# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the number of firms with market share
mstruct(share, type = "firm")
# Calculate top market share individually
top(share)
# Calculate the market structure group measures
mstruct(share, type = "all")
```
instead
```{r message=FALSE, warning=FALSE}
library(concstats)
# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the number of firms with market share
share_firm <- mstruct(share, type = "firm")
# Calculate top market share individually
share_top <- top(share)
# Calculate the market structure group measures
share_mstruct <- mstruct(share, type = "all")
```

In a Vignette 'Introduction to Concstats'

- 'Installation' paragraph is redundant since the vignette 'Getting started' covers that issue
- there is no code producing a plot titled 'Credit cooperatives (type) A'.
