### 2023-02-13 #559 @schneiderpy
G1.0 Statistical Software should list at least one primary reference from published academic literature.

In the list of references in the DESCRIPTION file I have not found Stenbacka index and Simpson. Please indicate where I can find it.

> The following REF has already been in  the DESCRIPTION file.
Melnik, A., Shy, Oz, Stenbacka, R., (2008), “Assessing market dominance”, Journal of Economic
Behavior and Organization 68, pp. 63-72.
I have just added Simpson

> G1.1 Statistical Software should document whether the algorithm(s) it implements are...

Not stated in README file. I guess it should be "The first implementation within R of an algorithm which has previously been implemented in other languages or contexts". I have seen a similar package in python.

I need some advise on this. The Guide states that it should be documented. The README file has a "Prior Art" section. Some functions (e.g. HHI) are already implemented in other packages. For others (e.g. dual of HHI, Stenbacka index .. ) it is a new implementation in R.
Since this is a R package I honestly did not check Python, but I also came accross the "concentrationMetrics" package in Python. Therefore I have added the following: "Some popular measures, e.g. Gini or the Herfindahl-Hirschman index have also
been implemented in Python."

> An argument unbiased explanation as 'Logical. Argument specifying whether or not a finite sample correction should be applied. The default is FALSE.' is misleading. Some measures can be normalized since their value - under equal shares assumption - depends on the number of market participants. Please correct the description of argument unbiased towards idea of normalization of the measure.

I have changed all "unbiased" to _normalized_ now. "specifying whether or not a normalized value is
#'  required. Ranges from {0, 1} and often used for comparison over time. Must
#'  be either TRUE or FALSE."

> There is an inconsistency between required form of 'x' vector. For entropy, shares are calculated from 'x' while for HH index, input is in a form of shares. It should be unified.

I have unified the required form to "x".

> Correct me, if I am wrong. Simpson index is defined in the package as a complement of HH index, that is
. Under equal shares assumption it gives. Following normalization of HH index, normalized Simpson index should be of the form ......

I need to check this. Furthermore, it should say "Gini-Simpson index"

> Function grs is implemented in a slightly different way then the way indicated in the reference. GRS does not to be restricted to positive shares. And so, sum(x > 0) should be replaced by length(x) or something similar. Such replacement has an impact on the result.

I have to check this too. However, in this context, negative shares make no sense ...

> Function dom has no normalization option. Correctly?

This is correct. The Dominnce index has no normalization option (removed this from the script). However, in theory, it is possible, since the dom index is based on the HHI.

> Man files
no citations included

All man files have now citations and references.

> for Stenbacka index, I have not found a relevant reference. Is it defined just for two biggest market shares?. I guess that one is fine: Melnik, A., Shy, Oz, Stenbacka, R., (2008), “Assessing market dominance”, Journal of Economic Behavior and Organization 68, pp. 63-72.

That's correct.

> he most of measures can be described briefly with single formula and understood by majority package users. I encourage the Author to enrich help pages with such description using TEX e.g. eqn and deqn environments. It looks nice :)

I need some advise on this point. I have "checked" some other packages, in particular stats packages. However, it seems that including formulae is not "best practice". Probably due to the references where one can "dive" deeper. I have no problem to include some well known formula, e.g. HHI index and others.

> description 'computes the numbers equivalent' of the function nrs_eq is very unclear.

Changed to: "\code{concstats_nrs_eq} computes the reciprocal of the HHI, which indicates the equivalent number of firms of the same size"

> in my opinion examples should be written to print outputs and informative, that is ...
changed to this form (the results are not stored in a named object)
library(concstats)
# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the number of firms with market share
mstruct(share, type = "firm")
# Calculate top market share individually
top(share)
# Calculate the market structure group measures
mstruct(share, type = "all")

> Vignettes
In a Vignette 'Introduction to Concstats'

'Installation' paragraph is redundant since the vignette 'Getting started' covers that issue

That is true. I have removed the "installation" from the "Intro" vignette, and kept it in the "Getting started" vignette.

> here is no code producing a plot titled 'Credit cooperatives (type) A'.

The vignette now shows the code of the plot.
