+++ DESCRIPTION
@@ -23,14 +23,17 @@ Description: Based on individual market shares of all participants in a
     analisis de la estructura de los mercados".
     4.  Ginevicius, R. and S. Cirba (2009). "Additive measurement of market
     concentration" <doi:10.3846/1611-1699.2009.10.191-198>.
-    5.  Herfindahl, O. C. (1950), "Concentration in the steel industry
-    (PhD thesis)".
-    6.  Hirschmann, A. O. (1945), "National power and structure of foreign trade".
+    5.  Herfindahl, O. C. (1950), "Concentration in the steel industry"
+    (PhD thesis).
+    6.  Hirschmann, A. O. (1945), "National power and structure of foreign
+    trade".
     7.  Melnik, A., O. Shy, and R. Stenbacka (2008), "Assessing market dominance"
     <doi:10.1016/j.jebo.2008.03.010>.
     8.  Palma, J. G. (2006). "Globalizing Inequality: 'Centrifugal' and
     'Centripetal' Forces at Work".
     9.  Shannon, C. E. (1948). "A Mathematical Theory of Communication".
+    10.  Simpson, E. H. (1949). "Measurement of Diversity"
+    <doi:10.1038/163688a0>
 License: MIT + file LICENSE
 URL: https://github.com/schneiderpy/concstats,
     https//schneiderpy.github.io/constats (website)
