+++ R/inequ.R
@@ -5,14 +5,15 @@
 #' @description A set of different inequality and diversity measures.
 #'
 #' @usage
-#'  concstats_inequ(x, unbiased = FALSE, type = c("entropy", "gini", "simpson",
-#'  "palma", "grs", "all"), na.rm = TRUE)
+#'  concstats_inequ(x, normalized = FALSE, type = c("entropy", "gini",
+#'  "simpson", "palma", "grs", "all"), na.rm = TRUE)
 #' @srrstats {G2.0a, G2.1a, EA1.1, EA1.3} accepted as input, length and type
 #' @param x a numeric vector of length 1 with non-negative values.
-#' @param unbiased Logical. Argument of the functions \code{concstats_entropy},
-#'  \code{concstats_gini}, and \code{concstats_simpson} specifying whether or
-#'  not a finite sample correction should be applied. Must be either TRUE or
-#'  FALSE. The default is FALSE.
+#' @param normalized Logical. Argument of the functions
+#'  \code{concstats_entropy}, \code{concstats_gini}, and
+#'  \code{concstats_simpson} specifying whether or not a normalized value is
+#'  required. Ranges from {0, 1} and often used for comparison over time. Must
+#'  be either TRUE or FALSE. The default is FALSE.
 #' @param type a character string of the measure to be calculated, defaults to
 #'  "concstats_entropy". Input is not case-sensitive.
 #' @param na.rm a logical vector that indicates whether \code{NA} values should
@@ -25,36 +26,51 @@
 #'  \code{concstats_palma}, \code{concstats_grs}, \code{concstats_all}.
 #'   If no measure is specified, "concstats_entropy" is the default.
 #'
-#'  \code{concstats_entropy} returns the (Shannon) Entropy,
-#'   \code{concstats_gini} is the Gini coefficient, \code{concstats_simpson} is
-#'    the complement of the Herfindahl-Hirschman Index. You can normalize each
-#'     of these three measures by setting \code{unbiased = TRUE}
+#'  \code{concstats_entropy} returns the Shannon Entropy (Shannon, 1948),
+#'  \code{concstats_gini} is the Gini coefficient,
+#'  \code{concstats_simpson} is the Gini-Simpson index (Simpson, 1949;
+#'   Jost, 2006), also known as the Gini impurity (Gini's diversity index) in
+#'   Machine Learning, Gibbs-Martin index or Blau index in sociology and
+#'   management studies. You can normalize each of these three measures by
+#'   setting \code{normalized = TRUE}
 #'  \code{concstats_palma} measures the ratio of inequality (normally used with
-#'   income inequality) of the top 10 percent to the bottom 40 percent.
-#'  \code{concstats_grs} is an alternative inequality
-#'   measure (Ginevicius, 2009),
+#'   income inequality) of the top 10 percent to the bottom 40 percent
+#'   (Palma, 2006).
+#'  \code{concstats_grs} is an alternative inequality measure (Ginevicius, 2009)
+#'   and
 #'  \code{concstats_all} returns all measures in a one step procedure.
 #'
 #' @return the calculated numeric measure
 #' @note the non-negative vector of market shares should be in a decimal form
 #'  corresponding to the total shares of individual firms/units.The vector
 #'  should sum up to 1.
+#' @references Shannon, C. E. (1948). A Mathematical Theory of Communication,
[…]
