+++ R/comp.R
@@ -5,14 +5,13 @@
 #' @description A set of different concentration measures.
 #'
 #' @usage
-#'  concstats_comp(x, unbiased = FALSE, type = c("hhi", "hhi_d", "hhi_min",
+#'  concstats_comp(x, normalized = FALSE, type = c("hhi", "hhi_d", "hhi_min",
 #'  "dom", "sten", "all"),na.rm = TRUE)
 #' @srrstats {G2.0a, G2.1a, EA1.1, EA1.3} accepted as input, length and type
 #' @param x A numeric vector of length 1 with non-negative values.
-#' @param unbiased Logical. Argument specifying whether or not a finite sample
-#'  correction should be applied. Must be either TRUE or FALSE. The default is
-#'  FALSE.
-#
+#' @param normalized Logical. Argument specifying whether or not a normalized
+#'  value is required. Ranges from {0, 1} and often used for comparison over
+#'  time. Must be either TRUE or FALSE. The default is FALSE.
 #' @param type A character string of the measure to be calculated,
 #'  can be abbreviated with the first letter. Defaults to "hhi". Input is not
 #'  case-sensitive.
@@ -26,11 +25,14 @@
 #'  \code{concstats_dom}, \code{concstats_sten}, \code{concstats_all}.
 #'  If no measure is specified "hhi" will be the default.
 #'
-#'  \code{concstats_hhi}, \code{concstats_hhi_min}, \code{concstats_hhi_d}
-#'  calculate the Herfindahl-Hirschman index, its minimum, and its dual,
-#'  respectively. \code{concstats_dom} calculates a dominance index and
-#'  \code{concstats_sten} calculates the Stenbacka index. The index indicates
-#'  the market share of a dominant position.
+#'  \code{concstats_hhi} calculates the widely used Herfindahl-Hirschman Index
+#'  (Herfindahl, 1950 and Hirschman, 1945), \code{concstats_hhi_min}
+#'  its minimum, and \code{concstats_hhi_d} its dual (Chang, Guerra,
+#'  de Souza Penaloza & Tabak, 2005)
+#'  \code{concstats_dom} calculates a dominance index (Garcia, 1994) and
+#'  \code{concstats_sten} calculates the Stenbacka index (Melnik, Shy, and
+#'  Stenbacka, 2005). The index indicates the market share of a dominant
+#'  position.
 #'
 #'  All measures can be accessed individually.
 #'  \code{concstats_hhi}, \code{concstats_hhi_d}, and \code{concstats_dom} can
@@ -40,25 +42,36 @@
 #'
 #' @return the calculated numeric measure.
 #' @note The vector of market shares should be in a decimal form corresponding
-#'  to total shares of individual firms/units. The sum of the vector should sum
-#'  up to 1.
+#'  to total shares of individual firms/units. The vector should sum up to 1.
+#' @references Herfindahl, O. C. (1950), "Concentration in the steel industry"
+#'  (PhD thesis)
+#' @references Hirschmann, A. O. (1945), "National power and structure of
+#'  foreign trade".
+#' @references Chang, E. J., Guerra, S. M., de Souza Penaloza, R. A. & Tabak,
+#'  B. M. (2005) Banking concentration: the Brazilian case. In Financial
+#'  Stability Report. Brasilia: Banco Central do Brasil, 4: 109-129.
+#' @references Garcia Alba Idunate, P. (1994). Un Indice de dominancia para el
+#'  analisis de la estructura de los mercados. El Trimestre Economico,
[…]
