### 2023-02-13 98a5b52
Address reviewer comments (Issue #559) added refs to manual
