### 2023-02-13 33acd50
Address reviewer comments (Issue # 559) .. added code for plot
