### 2023-02-14 #559 @SebastianWojcik86
@schneiderpy implemented a majority of my suggests, minority of my suggests is under consideration or revision. I am very glad that my review was taken into account in such scope.
