+++ man/concstats_concstats.Rd
@@ -41,7 +41,7 @@ to the total share of individual firms/units. The vector should sum up to 1.
 
 }
 \seealso{
-\code{\link[concstats_mstruct]{concstats_mstruct()}},
-\code{\link[concstats_comp]{concstats_comp()}},
-\code{\link[concstats_inequ]{concstats_inequ()}}
+\code{\link[=concstats_mstruct]{concstats_mstruct()}},
+\code{\link[=concstats_comp]{concstats_comp()}},
+\code{\link[=concstats_inequ]{concstats_inequ()}}
 }
