### 2023-02-14 3b00b69
Address reviewer comments (Issue #559)
