+++ tests/testthat/test_inequ.R
@@ -19,19 +19,19 @@ test_that("concstats_inequ function operates / switches properly", {
 
   expect_vector(x, ptype = numeric(), size = 4)
   expect_equal(concstats_inequ(x, type = "entropy"),
-               concstats_entropy(x, unbiased = FALSE))
-  expect_equal(concstats_inequ(x, unbiased = TRUE, type = "entropy"),
-               concstats_entropy(x, unbiased = TRUE))
+               concstats_entropy(x, normalized = FALSE))
+  expect_equal(concstats_inequ(x, normalized = TRUE, type = "entropy"),
+               concstats_entropy(x, normalized = TRUE))
   expect_equal(concstats_inequ(x1, type = "gini"), concstats_gini(x1))
-  expect_equal(concstats_inequ(x1, unbiased = TRUE, type = "gini"),
-               concstats_gini(x1, unbiased = TRUE))
+  expect_equal(concstats_inequ(x1, normalized = TRUE, type = "gini"),
+               concstats_gini(x1, normalized = TRUE))
   expect_equal(concstats_inequ(x, type = "simpson"), concstats_simpson(x))
-  expect_equal(concstats_inequ(x, unbiased = TRUE, type = "simpson"),
-               concstats_simpson(x, unbiased = TRUE))
+  expect_equal(concstats_inequ(x, normalized = TRUE, type = "simpson"),
+               concstats_simpson(x, normalized = TRUE))
   expect_equal(concstats_inequ(x1, type = "palma"), concstats_palma(x1))
   expect_equal(concstats_inequ(x2, type = "grs"), concstats_grs(x2))
   expect_equal(concstats_inequ(x, type = "Entropy"), concstats_entropy(x,
-                                                            unbiased = FALSE))
+                                                            normalized = FALSE))
   expect_equal(concstats_inequ(x, type = "all"), concstats_all_inequ(x))
   expect_error(concstats_inequ(x1b, na.rm = TRUE))
 })
@@ -53,7 +53,7 @@ test_that("concstats_entropy function operates properly", {
   x9 <- c(NA, NA, NA, NA, NA, NA, NA, NA)
   xch <- c("a", "b", "c", "d", "e", "f", "g", "h")
   na.rm <- as.logical(TRUE | FALSE)
-  unbiased <- as.logical(TRUE | FALSE)
+  normalized <- as.logical(TRUE | FALSE)
 
 #' @srrstats {G5.3} Expected to return objects containing no missing (`NA`)
   expect_true(any(is.na(x2)), all(!is.na(x2)))
@@ -72,7 +72,7 @@ test_that("concstats_entropy function operates properly", {
   expect_error(concstats_entropy(x8, na.rm = TRUE))
   expect_error(concstats_entropy(x1b, na.rm = TRUE))
   expect_warning(concstats_entropy(x2, na.rm = 0))
-  expect_warning(concstats_entropy(x2, unbiased = 0))
+  expect_warning(concstats_entropy(x2, normalized = 0))
 #' @srrstats {G3.0, EA6.0, EA6.0e} Return values, single-valued objects.
   expect_equal(concstats_entropy(x, isTRUE(all.equal(1, sum(x),
                                        tolerance = .Machine$double.eps^0.25))),
@@ -88,8 +88,8 @@ test_that("concstats_entropy function operates properly", {
                "vector x in `concstats_entropy` does not sum to 1")
   expect_length(na.rm, 1L)
   expect_type(na.rm, "logical")
-  expect_length(unbiased, 1L)
-  expect_type(unbiased, "logical")
[…]
