### 2023-02-15 #559 @schneiderpy
> Correct me, if I am wrong. Simpson index is defined in the package as a complement of HH index, that is

. Under equal shares assumption it gives. Following normalization of HH index, normalized Simpson index should be of the form ......

I need to check this. Furthermore, it should say "Gini-Simpson index"

This issue has been addressed now using the "Gini-Simpson index" also known as Gini's diversity index, or Gini impurity in ML. Since this index ranges from {0, 1} there is no normalization option included.
