### 2023-02-15 123d7fe
Addressed reviewer comments (Issue # 559), Simpson index to Gini-Simpson index
