### 2023-02-23 ffbd8cc
removed equation
