### 2023-02-23 fc1a486
updated equation for grs measure
