### 2023-02-23 #559 @schneiderpy
Hello @SebastianWojcik86

> Function grs is implemented in a slightly different way then the way indicated in the reference. GRS does not to be restricted to positive shares. And so, sum(x > 0) should be replaced by length(x) or something similar. Such replacement has an impact on the result.

The equation for the "grs measure" has been updated to
grs <- as.numeric(sum((length(x) ^ 2 * x[1] + 0.3 * x ^ 2) /
                      (length(x) ^ 2 + length(x) * 0.3 * x[1] * x) * x))
This is to get the implementation of the equation right. However, the implementation of length(x) has no impact on the results, since, in this context, negative shares make no sense, and therefore the input is still restricted to x > 0.

With that said, I have replied to all comments of the reviewer @SebastianWojcik86. Most of them have been implemented to improve the package. Again, thank you @SebastianWojcik86
