### 2023-02-24 #559 @SebastianWojcik86
Thank you @schneiderpy for taking my suggestions into account. I fully accept a current version of the package.
