### 2023-02-25 #559 @christopherkenny
Thank you for the opportunity to review this package. It is well designed and fairly comprehensive. I have many small suggestions and requests for clarification within documentation are detailed below. More importantly, I am concerned about the state of the tests used. Substantial revision is necessary to the `testthat` tests to ensure that the tests test functions from the package.

---

### Compliance with Standards

- [x] This package complies with a sufficient number of standards for a **silver** badge
- [x] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: None.

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways:

- *6.1.1*: The package is extensively documented, including detailed citations and extensive comparisons with other softwares implementing related and overlapping measures. Functions are documented using `roxygen2` and are detailed. Only minor requests are suggested re documentation.
- *6.1.2*: The package checks inputs across all functions, checking types and lengths. It uses appropriate calls to `match.arg()` and handles missing values transparently. Note that the package does not accept tabular inputs, so this is constrained to extensive vector/scalar usage. Only minor requests are suggested for input checks and behaviors.
- *6.1.5*: For silver, only two categories are necessary. Compliance appears insufficient here, pending revisions.

Absent test revision, the standard needed for silver appears to be already met, from my reading of the instructions.

---

### General Review

There seem to be two sets of reviewer instructions, [one in the devguide](https://devguide.ropensci.org/reviewtemplate.html#reviewtemplate) and [one in the linked instructions](https://stats-devguide.ropensci.org/pkgreview.html) from this issue. I first go through the check-boxes from the devguide and then more extensive comments follow using the linked instructions.

#### Functionality (devguide)

    - Detailed issues with this are explained below. All tests do pass, but some of the tests do not actually test the package.

- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
     - It may be worth linking to `CONTRIBUTING.md` from `README.md`, as you do mention development briefly in the readme.
- [x] The documentation is sufficient to enable general use of the package beyond one specific use case

- Throughout the package (documentation, readme, and vignettes), you opt not to document functions following standard practices as `fn()`, but instead as `fn`. This is a fine discretionary choice and does not appear to be part of rOpenSci's guidelines for review. However, you do use `pkgdown` and `roxygen2`, which makes this a surprising choice, as the functions do not link to the documentation when you do not use `fn()` notation. It would be useful for users of the package if you allow for the automatic links to documentation to generate by using `fn()` over `fn`.
- You often make use of `@rdname` for related families of functions with overlapping applications. This looks well done, but there is a bit of repetition here. You re-document several `@param`s on the same `@rdname`. For consistency here, the duplicated `@param`s should likely be removed or the `@rdname` tag should be removed in favor of separate documentation pagess.
- In `concstats_comp()` and `concstats_inequ()`, `normalize` should be further documented and potentially checked.
    - If `normalize = NA`, it gives a generic R warning. It seems that this should be checked along with the logical and length checks, as you are already checking those and providing a directly helpful error message.
    - Further, `normalize` is summarily ignored if `type = 'all'`. This may be intended behavior, but should be documented or checked for. Otherwise, it confidently gives the wrong (or at least unexpected) answer when `normalize = TRUE` for the gini and entropy measures compared to when `normalize = TRUE` and `type %in% c('gini', 'entropy')` or `type %in% c('hhi')`
- Line 40 of `R/comp.R` references a function (`concstats_all()`) which does not exist. This should be revised to point to (presumably)`concstats_all_comp()`.
- Line 39 of `R/inequ.R` references a function (`concstats_all()`) which does not exist. This should be revised to point to (presumably)`concstats_all_inequ()`.
- Line 27 of `R/mstruct.R` references a function (`concstats_all()`) which does not exist. This should be revised to point to (presumably)`concstats_all_mstruct()`.
- Documentation for `concstats_concstats()` would benefit from the list of measures that will be computed. It indicates "a set of different and selected structural, inequality, and concentration measures" will be computed, but indicating which those are would be useful to users.
- `x` is documented as "@param x a numeric vector of length 1 with non-negative values." throughout `R/comp.R`, `R/concstats.R`, `R/inequ.R`, and `R/mstruct.R`. This seems incorrect and conflicts with the examples, which _always_ show a numeric vector of positive length.
- When documenting `@param na.rm`, the following is repeated "If set to \code{FALSE} the computation yields \code{NA}.". However, this is not true. It returns `NA` only if there is an `NA` value in the vector.
- The package relies on `readr`, but does not use `readr` internally (only in `data-raw`, which does not then require it be imported). This should be removed from the DESCRIPTION file and the `data.R` roxygen.
- `dplyr` should not be imported (for the same reasons as `readr`) and should be removed from DESCRIPTION and the `data.R` roxygen, but should be `suggests` for its use in the vignettes.

#### Algorithms

- The package is well-written and most choices make sense. The package appropriately uses vectorized computation and avoids repetition by pointing to shared functions where necessary. This should scale appropriately for most applications.
- Input classes and lengths are consistently checked. One point of question: When checking `x`, what is the intended behavior if `length(x) == 1`? In some cases, this works. In others, such as for `concstats_grs()`, it fails with an uninformative error.
- `concstats_constats()`, `concstats_all_comp()`, `concstats_all_inequ()`, and `concstats_all_mstruct()` each round the output to 2 decimal digits. This may be a reasonable choice, but it is not clear why 2 decimal digits is appropriate here, but not in the other functions. Should all outputs contain a rounding step or is this step unnecessary? One way to make this more consistent could be to make this optional, controlled by a new argument.
- In each of  `concstats_all_comp()`, `concstats_all_inequ()`, and `concstats_all_mstruct()`, there is code which runs via `invisible(utils::capture.output(...))`.  It is not clear what the purpose of this is, as there does not appear to be any print statements in the functions in the `...`. If this is intended to capture some known issue with printing elsewhere, it would seem best to deal with this directly by modifying inputs or providing warnings or errors where appropriate.

- The tests here are quite extensive, indeed there are 446 tests using `testthat`. Coverage is fair, including many tests for things like stray `NA`s or incorrect input types. The tests are extensively documented. There are frequent tests for correctness of calculations and references for the sources of correct values are largely included.
- However, a concerning number of the tests in the package are inappropriate for testing the package. Hundreds of the tests are testing basic aspects of base R, not testing pieces of the package. Things like testing R's assignment operator, that NAs are NA, or that vectors sum properly. These tests should be removed.
- In many calls to `expect_error()` across all test files it appears that the parentheses are frequently not matched properly, so that what appears to be an argument for `regexp` is instead passed to `isTRUE()` or to `concstats_*()`. That will always cause an error, so the test is again not testing aspects of the package, but instead that a base R function performs how it is intended. In many cases, this means that tests designated for particular standards with "@srrstats {G5.2, G5.2a, G5.2b, G5.8, G5.8a, G5.8b}" (e.g. in `tests/testthat/test_comp.R` l. 64-68) do not appear to test anything that would satisfy those standards for `concstats`. This is a pervasive problem in the tests that would need substantial revision.

#### Package Design

- The categorization of functions here is well-documented and the options between a character `type` parameter or a specific function for any particular measure give users multiple ways to use the package. This offers a friendly interface to many different measures.
- Inputs are all consistenly designed to take in vectors or scalars. This seems appropriate and sufficiently friendly for broad use. While other types of inputs could be considered (matrices, for example), the current set up seems like the right choice, as the functions are fast and simple, so that these functions could be mapped over other types of inputs in a simple way by users.
- In `data-raw/creditcoops.R`, on line 3 `n()` should be `dplyr::n()`, otherwise the demo data cannot be reproduced (assuming no packages loaded beyond `devtools::load_all()`).

---
