### 2023-02-25 #559 @schneiderpy
Thank you @christopherkenny for reviewing this package with a lot of new input and (still) things to do. Some of these issues already poped up earlier (e.g. dependencies) others are new, and probably due to inexperience. Nevertheless, I'll do my best to address your issues and requests as soon as possible.
