### 2023-02-27 #559 @Paula-Moraga
Many thanks @SebastianWojcik86 and @christopherkenny for your time and effort to review the package!

Many thanks @schneiderpy for addressing the comments by @SebastianWojcik86. Looking forward to see the new version implementing the suggestions by @christopherkenny.
