### 2023-03-13 b2bccb2
changed to v. 0.1.6; addressed reviewer comments (Issue #559)
