+++ README.md
@@ -36,13 +36,14 @@ important to determine public policies and strategic corporate
 decisions. However, in research and in practice the most commonly used
 measure is the Herfindahl Hirschman Index. Various complementary or
 alternative measures exist, which - used as a set - might reduce
-uncertainty. The goal of the `concstats` package is to offer a set of
-alternative and/or additional measures for researchers in social
-sciences and practitioners in institutions concerned with competition on
-a regular basis to better determine a given market structure and
-therefore reduce uncertainty with respect to a given market situation.
-Various functions or groups of functions are available to achieve the
-desired goal.
+uncertainty. The goal of the
+[concstats](https://cran.r-project.org/package=concstats) package is to
+offer a set of alternative and/or additional measures for researchers in
+social sciences and practitioners in institutions concerned with
+competition on a regular basis to better determine a given market
+structure and therefore reduce uncertainty with respect to a given
+market situation. Various functions or groups of functions are available
+to achieve the desired goal.
 
 \-`concstats_concstats` calculates a set of pre-selected concentration
 and diversity measures in a one-step procedure.  
@@ -118,7 +119,7 @@ x2 <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04) # market shares of each firm in
                                           # the market (should sum up to 1)
 
 ## Calculate a selected set of market structure and concentration measures
-concstats_concstats(x2) # calculates a selected set of measures
+concstats_concstats(x2, digits = 2) # calculates a selected set of measures
 #>         Measures Values
 #> 1          Firms   6.00
 #> 2 Nrs_equivalent   3.33
@@ -148,7 +149,7 @@ package.
     [hhi](https://joss.theoj.org/papers/10.21105/joss.00828) and the
     [divseg](https://github.com/christopherkenny/divseg) packages. While
     the *hhi* package has just one function, neither of both packages
-    offer a finite sample correction.
+    offer a normalized version of the measure.
 
 -   The latter offers as well functions for the Entropy, Gini and
     Simpson measures.
@@ -161,8 +162,8 @@ package.
 Some popular measures, e.g. Gini or the Herfindahl-Hirschman index have
 also been implemented in Python.
 
-However, almost none of these packages offer finite sample correction,
-with the exception of the
+However, almost none of these packages offer a normalized calculation of
+the respective measure, with the exception of the
 [ineq](https://cran.r-project.org/package=ineq) package. Other functions
 are new implementations in R, e.g. Dominance Index, Palma ratio,
 Stenbacka Index, GRS measure, and the dual of the Herfindahl Hirschman
@@ -171,26 +172,24 @@ Index.
 ## References
 
 1.  Chang, E. J., Guerra, S. M., de Souza Penaloza, R. A. & Tabak, B. M.
-    2005) Banking concentration: the Brazilian case. In Financial
+    2005) “Banking concentration: the Brazilian case” In Financial
           Stability Report. Brasilia: Banco Central do Brasil, 4:
           109-129.
-2.  Cobham, A. and A. Summer (2013). Is It All About the Tails? The
-    Palma Measure of Income Inequality, Center for Global Development,
+2.  Cobham, A. and A. Summer (2013). “Is It All About the Tails? The
+    Palma Measure of Income Inequality”, Center for Global Development,
     Washington, DC.
[…]
