+++ NAMESPACE
@@ -22,9 +22,3 @@ export(concstats_sten)
 export(concstats_top)
 export(concstats_top3)
 export(concstats_top5)
-importFrom(dplyr,"%>%")
-importFrom(dplyr,filter)
-importFrom(dplyr,mutate)
-importFrom(readr,read_csv)
-importFrom(utils,data)
-importFrom(utils,head)
