### 2023-03-13 3c82582
Addresses several reviewer comments (Issue #559)
