+++ R/concstats.R
@@ -3,61 +3,58 @@
 #' @srrstats {G1.4} roxygen2 used to document functions
 #' @srrstats {EA1.2} scope of the package/ kind of questions the software is
 #'   intended to solve.
-#' @description a convenience function which calculates a selected set of
+#' @description A convenience function which calculates a selected set of
 #'  different market structure, inequality and concentration measures more or
 #'  less commonly used, e.g. k-firm ratios, Entropy, HHI, Palma ratio,
 #'  and others in a one step procedure to provide a first overview.
 #'
-#' @usage concstats_concstats(x, na.rm = TRUE)
+#' @usage concstats_concstats(x, na.rm = TRUE, digits = NULL)
 #' @srrstats {G2.0, G2.0a, G2.1, G2.1a, EA1.1, EA1.3} accepted as input, length
 #'  and type
-#' @param x a numeric vector of length 1 with non-negative values.
-#' @param na.rm a logical vector that indicates whether \code{NA} values should
-#'  be excluded or not. Must be either TRUE or FALSE. Defaults to TRUE. If set
-#'  to \code{FALSE} the computation yields \code{NA}.
+#' @param x A non-negative numeric vector.
+#' @param na.rm A logical vector that indicates whether \code{NA} values should
+#'  be excluded or not. Must be either \code{TRUE} or \code{FALSE}. Defaults to
+#'  \code{TRUE}. If set to \code{FALSE} the computation yields \code{NA} if
+#'   vector contains \code{NA} values.
+#' @param digits A non-null value for digits specifies the minimum number of
+#'  significant digits to be printed in values. The default is \code{NULL} and
+#'  will use base R print option. Significant digits defaults to 7.
 #' @details \code{concstats_concstats} computes a set of different and selected
-#'  structural, inequality, and concentration measures in a one step procedure,
-#'  however, all measures can be computed individually or in groups.
+#'  structural, inequality, and concentration measures in a one step procedure.
+#'  The resulting `data frame` contains eight measures: number of firms with
+#'  market share, numbers equivalent, the cumulative share of the top
+#'  (top 3 and top 5) firm(s) in percentage, the hhi index, the entropy index,
+#'  and the palma ratio. However, all measures can be computed individually or
+#'  in groups.
 #'
-#' @return returns a data frame of numeric measures with default settings.
-#' .
-#' @note the vector of market shares should be in a decimal form corresponding
+#' @return A `data frame` of numeric measures with default settings.
+#'
+#' @note The vector of market shares should be in a decimal form corresponding
 #'  to the total share of individual firms/units. The vector should sum up to 1.
 #'
-#' @seealso \code{\link[=concstats_mstruct]{concstats_mstruct()}},
-#'  \code{\link[=concstats_comp]{concstats_comp()}},
-#'  \code{\link[=concstats_inequ]{concstats_inequ()}}
+#' @seealso [concstats_mstruct()], [concstats_comp()], [concstats_inequ()]
 #'
 #' @examples
 #' # a vector of market shares
-#' # share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
+#' x <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
 #' # a selected set of different structural, concentration, and inequality
 #' # measures
-#' # concstats_concstats(share)
+#' concstats_concstats(x, digits = 2)
 #'
 #' @export
 #' @srrstats {G2.6, EA2.6} process vector data
-concstats_concstats <- function(x, na.rm = TRUE) {
[…]
