+++ R/comp.R
@@ -1,4 +1,4 @@
-#' @title Concentration Measures
+#' @title Group of Concentration Measures
 #'
 #' @srrstats {G1.4} roxygen2 used to document functions
 #'
@@ -6,60 +6,39 @@
 #'
 #' @usage
 #'  concstats_comp(x, normalized = FALSE, type = c("hhi", "hhi_d", "hhi_min",
-#'  "dom", "sten", "all"),na.rm = TRUE)
+#'  "dom", "sten", "all"), na.rm = TRUE, digits = NULL)
 #' @srrstats {G2.0a, G2.1a, EA1.1, EA1.3} accepted as input, length and type
-#' @param x A numeric vector of length 1 with non-negative values.
+#' @param x A non-negative numeric vector.
 #' @param normalized Logical. Argument specifying whether or not a normalized
-#'  value is required. Ranges from {0, 1} and often used for comparison over
-#'  time. Must be either TRUE or FALSE. The default is FALSE.
+#'  value is required. Ranges from (0, 1) and often used for comparison over
+#'  time. Must be either \code{TRUE} or \code{FALSE}. The default
+#'  is \code{FALSE}.
 #' @param type A character string of the measure to be calculated,
 #'  can be abbreviated with the first letter. Defaults to "hhi". Input is not
 #'  case-sensitive.
-#' @param na.rm a logical vector that indicates whether \code{NA} values should
-#'  be excluded or not. If set to \code{FALSE} the computation yields \code{NA}.
-#'  Must be either TRUE or FALSE. Defaults to TRUE.
-#'
+#' @param na.rm A logical vector that indicates whether \code{NA} values should
+#'  be excluded or not. If set to \code{FALSE} the computation yields \code{NA}
+#'  if vector contains \code{NA} values.
+#'  Must be either \code{TRUE} or \code{FALSE}. Defaults to \code{TRUE}.
+#' @param digits A non-null value for digits specifies the minimum number of
+#'  significant digits to be printed in values. The default is \code{NULL} and
+#'  will use base R print option.  Significant digits defaults to 7.
 #' @details
 #'  \code{concstats_comp} is a wrapper for the proposed concentration measures
-#'  \code{concstats_hhi}, \code{concstats_hhi_d}, \code{concstats_hhi_min},
-#'  \code{concstats_dom}, \code{concstats_sten}, \code{concstats_all}.
-#'  If no measure is specified "hhi" will be the default.
+#'  [concstats_hhi()], [concstats_hhi_d()], [concstats_dom()],
+#'  [concstats_hhi_min()], [concstats_sten()], [concstats_all_comp()].
+#'   If no measure is specified "hhi" will be the default.
+#'  \code{concstats_hhi}, can be calculated individually as a normalized
+#'   measure changing the default setting to \code{TRUE}.
+#'  \code{concstats_all_comp} computes all measures in a one step procedure.
+#'   For more details or references please see the help page of the respective
+#'   function.
 #'
-#'  \code{concstats_hhi} calculates the widely used Herfindahl-Hirschman Index
-#'  (Herfindahl, 1950 and Hirschman, 1945),
-#'  \code{concstats_hhi_min} its minimum, and \code{concstats_hhi_d} its dual
-#'  (Chang, Guerra, de Souza Penaloza & Tabak, 2005)
-#'  \code{concstats_dom} calculates a dominance index (Garcia, 1994) and
[…]

[…]
