### 2023-03-13 f9cdfa6
Addressed reviewer comments (Issue #559); updated documentation
