### 2023-03-13 02e6e07
Addressed reviewer comments (Issue #559) expanded documentation
