+++ tests/testthat/test_comp.R
@@ -13,8 +13,6 @@ test_that("concstats_comp function operates / switches properly", {
   x8 <- c(-0.2, -0.3, -0.4, -0.100001)
   x9 <- c(NA, NA, NA, NA, NA, NA, NA, NA)
   xch <- c("a", "b", "c", "d", "e", "f", "g", "h")
-  na.rm <- as.logical(TRUE, FALSE)
-  normalized <- as.logical(TRUE, FALSE)
 
 #' @srrstats {G5.3} Expected to return objects containing no missing (`NA`)
   expect_true(any(is.na(x2)), all(!is.na(x2)))
@@ -31,8 +29,7 @@ test_that("concstats_comp function operates / switches properly", {
   expect_equal(concstats_comp(x, type = "HHI"), concstats_hhi(x))
   expect_equal(concstats_comp(x, type = "ALL"), concstats_all_comp(x))
   expect_error(concstats_comp(x1b, na.rm = TRUE))
-  expect_length(normalized, 1L)
-  expect_type(normalized, "logical")
+
 })
 
 ## concstats_hhi
@@ -46,42 +43,38 @@ test_that("concstats_hhi function operates properly", {
   x2 <- c(0.2, 0.3, 0.4, 0.1, NA)
   x3 <- c(0.2, 0.4, 0.15, 0.05, 0.22, 0.9)
   x4 <- c(20, 30, 40, 10)
-  x5 <- c(0.2, 0.3, 0.4, 0.101)
-  x6 <- c(0.2, 0.3, 0.4, 0.1, 0.0001)
+  x5 <- c(0.2, 0.3, 0.4, 0.10001)
   x8 <- c(-0.2, -0.3, -0.4, -0.100001)
   x9 <- c(NA, NA, NA, NA, NA, NA, NA, NA)
   xch <- c("a", "b", "c", "d", "e", "f", "g", "h")
-  na.rm <- as.logical(TRUE | FALSE)
-  normalized <- as.logical(TRUE | FALSE)
 
 #' @srrstats {G5.3} Expected to return objects containing no missing (`NA`)
   expect_true(any(is.na(x2)), all(!is.na(x2)))
 
   expect_true(all(round(x) == 0), (abs(x) > 0 & abs(x) <= 1))
   expect_vector(x, ptype = double(), size = 4)
-  expect_equal(x, as.numeric(x4 / sum(x4)))
   expect_equal(concstats_hhi(x2, na.rm = FALSE), NA_real_)
 #' @srrstats {G5.2, G5.2a, G5.2b, G5.8, G5.8a, G5.8b} Edge test for data of
 #'  unsupported types
-  expect_error(concstats_hhi(xch, !isTRUE(is.numeric(xch),
-  "x in `concstats_hhi` must be a numeric vector\n",
-  "You have provided an object of class: ", class(x)[1])))
+  expect_error(concstats_hhi(xch, !is.numeric(xch)))
 #' @srrstats {G 5.0, G5.2, G5.2a, G5.2b, G5.8c} Error and error messages on
 #'  vector with all-`NA` fields
   expect_error(concstats_hhi(x9, na.rm = TRUE))
   expect_error(concstats_hhi(x8, na.rm = TRUE))
   expect_error(concstats_hhi(x1b, na.rm = TRUE))
-  expect_warning(concstats_hhi(x, na.rm = 0))
-  expect_warning(concstats_hhi(x, normalized = 0))
+  expect_error(concstats_hhi(x, na.rm = 0))
+  expect_error(concstats_hhi(x, normalized = 0))
+#' @srrstats {G3.0, EA6.0, EA6.0e} Return values single-valued objects.
[…]
