+++ tests/testthat/test_inequ.R
@@ -22,9 +22,10 @@ test_that("concstats_inequ function operates / switches properly", {
                concstats_entropy(x, normalized = FALSE))
   expect_equal(concstats_inequ(x, normalized = TRUE, type = "entropy"),
                concstats_entropy(x, normalized = TRUE))
-  expect_equal(concstats_inequ(x1, type = "gini"), concstats_gini(x1))
+  expect_equal(concstats_inequ(x1, type = "gini"),
+               concstats_gini(x1, normalized = FALSE))
   expect_equal(concstats_inequ(x1, normalized = TRUE, type = "gini"),
-               concstats_gini(x1, normalized = TRUE))
+               concstats_gini(x1))
   expect_equal(concstats_inequ(x, type = "simpson"), concstats_simpson(x))
   expect_equal(concstats_inequ(x1, type = "palma"), concstats_palma(x1))
   expect_equal(concstats_inequ(x2, type = "grs"), concstats_grs(x2))
@@ -32,6 +33,9 @@ test_that("concstats_inequ function operates / switches properly", {
                                                             normalized = FALSE))
   expect_equal(concstats_inequ(x, type = "all"), concstats_all_inequ(x))
   expect_error(concstats_inequ(x1b, na.rm = TRUE))
+  expect_error(concstats_inequ(x, na.rm = 0))
+  expect_error(concstats_inequ(x, normalized = NA))
+
 })
 
 ## concstats_entropy
@@ -45,49 +49,39 @@ test_that("concstats_entropy function operates properly", {
   x2 <- c(0.2, 0.3, 0.4, 0.1, NA)
   x3 <- c(0.4, 0.3, 0.2, 0.1)
   x4 <- c(20, 30, 40, 10)
-  x5 <- c(0.2, 0.3, 0.4, 0.101)
+  x5 <- c(0.2, 0.3, 0.4, 0.1000001)
   x6 <- c(0.2, 0.3, 0.4, 0.1, 0.0001)
   x8 <- c(-0.2, -0.3, -0.4, -0.100001)
   x9 <- c(NA, NA, NA, NA, NA, NA, NA, NA)
   xch <- c("a", "b", "c", "d", "e", "f", "g", "h")
-  na.rm <- as.logical(TRUE | FALSE)
-  normalized <- as.logical(TRUE | FALSE)
 
 #' @srrstats {G5.3} Expected to return objects containing no missing (`NA`)
   expect_true(any(is.na(x2)), all(!is.na(x2)))
 
   expect_true(all(round(x) == 0), (abs(x) > 0 & abs(x) <= 1))
   expect_vector(x, ptype = numeric(), size = 4)
-  expect_equal(x, as.numeric(x4 / sum(x4)))
   expect_equal(concstats_entropy(x2, na.rm = FALSE), NA_real_)
 #' @srrstats {G5.2, G5.2a, G5.2b, G5.8, G5.8a, G5.8b} Edge test for data of
 #'  unsupported types
-  expect_error(concstats_entropy(xch, !isTRUE(is.numeric(xch),
-  "x in `concstats_entropy` must be a numeric vector\n",
-  "You have provided an object of class: ", class(x)[1])))
+  expect_error(concstats_entropy(xch, !is.numeric(xch)))
 #' @srrstats {G5.2, G5.2a, G5.2b, G5.8c} Error on vector with all-`NA` fields
   expect_error(concstats_entropy(x9, na.rm = TRUE))
   expect_error(concstats_entropy(x8, na.rm = TRUE))
   expect_error(concstats_entropy(x1b, na.rm = TRUE))
-  expect_warning(concstats_entropy(x2, na.rm = 0))
[…]

[…]
