### 2023-03-13 b8459bb
Addressed reviewer comments (Issue #559); updated documentation
