+++ vignettes/concstats-intro.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Introduction to concstats"
 author: "Andreas Schneider"
-date: "`r Sys.Date()`"
+date: "Last updated `r Sys.Date()`"
 output: 
   rmarkdown::html_vignette:
     toc: TRUE
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(collapse = TRUE,
 ```
 
 
-The goal of the ``concstats`` package is to offer a set of alternative and/or
+The goal of the `concstats` package is to offer a set of alternative and/or
 additional measures to better determine a given market structure and therefore
 reduce uncertainty with respect to a given market situation. Various functions
 or groups of functions are available to achieve the desired goal. 
@@ -43,7 +43,10 @@ are 22 paired observations for real Paraguayan credit cooperatives
 (with assets > 11 Mio. USD) for 2016 and 2018 with their respective total loans 
 granted. For a better visualization there is an additional column with the 
 transformed total loans. For further information on the data please see the
-`data` help file.
+`creditcoops`
+[help](https://schneiderpy.github.io/concstats/reference/creditcoops.html)
+file. For a practical implementation you might be interested
+to read the following [article](https://doi.org/10.1515/zfgg-2022-0002).
 
 ```{r}
 data("creditcoops")
@@ -63,10 +66,10 @@ head(creditcoops)
 
 ## Group wrapper overview
 At the moment, there are the following groups of functions available:    
-- `concstats_mstruct` is a wrapper for market structure measures    
-- `concstats_inequ` is a wrapper for inequality and diversity measures    
-- `concstats_comp` is a wrapper with different concentration measures    
-- `concstats_concstats` is a function which calculates a set of pre-selected
+- `concstats_mstruct()` is a wrapper for market structure measures    
+- `concstats_inequ()` is a wrapper for inequality and diversity measures    
+- `concstats_comp()` is a wrapper with different concentration measures    
+- `concstats_concstats()` is a function which calculates a set of pre-selected
                       measures in a one step procedure to get a quick overview
                       of a given market structure    
 
@@ -78,19 +81,24 @@ We will use a vector which represents market participants with their respective
 market shares (in decimal form): 
 
 ### concstats_mstruct
-The wrapper includes the following individual functions:
-```firm```, ```nrs_eq```, ```top```, ```top3```, ```top5```, and ```all```.
-All individual functions can be accessed directly with the prefix "concstats_",
-e.g. "concstats_firm".
+The wrapper includes the following arguments to calculate individual functions:
+``firm``, ``nrs_eq``, ``top``, ``top3``, ``top5``, and ``all``. You can also
+pass the additional argument ``digits`` to ``all``, which controls for the
+number of digits to be printed.  
+All individual functions can be accessed directly with the prefix "concstats_"
+(e.g. "concstats_firm" or "concstats_all_mstruct").
 ```{r example-1}
 test_share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04, 0, 0)
 test_share_top5 <- concstats_top5(test_share) # top 5 market share in percentage
 test_share_top5
 ```
 
-You should have noticed that the market shares are in decimal form. There are 
-eight market participants, however, two have no market shares, by default 
-```concstats``` treats 0 as NA. The result is a top 5 market share of 96 %.
+You should have noticed that the market shares are in decimal form. However, you
+can use integers or floating point numeric types to express market shares. All
+`concstats` functions will take care of this and convert theses vectors into
[…]
