### 2023-03-13 #559 @schneiderpy
@christopherkenny, apologize for the delay.
 > Automated tests: Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
> Detailed issues with this are explained below. All tests do pass, but some of the tests do not actually test the package.

Response: That was true. for more details see below.

>  Community guidelines including contribution guidelines in the README or CONTRIBUTING

It may be worth linking to CONTRIBUTING.md from README.md, as you do mention development briefly in the readme.

Response: There is now a link from README to CONTRIBUTING.md

> Throughout the package (documentation, readme, and vignettes), you opt not to document functions following standard practices as fn(), but instead as fn. This is a fine discretionary choice and does not appear to be part of rOpenSci's guidelines for review. However, you do use pkgdown and roxygen2, which makes this a surprising choice, as the functions do not link to the documentation when you do not use fn() notation. It would be useful for users of the package if you allow for the automatic links to documentation to generate by using fn() over fn.

Response: I have changed this according to standard practice. All fs() have now links in the documentation files to make use of the pkgdown and roxygen2 functionality to create autmatic links.

>  You often make use of @rdname for related families of functions with overlapping applications. This looks well done, but there is a bit of repetition here. You re-document several @params on the same @rdname. For consistency here, the duplicated @params should likely be removed or the @rdname tag should be removed in favor of separate documentation pages

Response: I have considered your suggestion. All @rdname tags have now the name acording to the function, that is, all functions have now their own documentation page. I did this is favor of more readability and it gives me more flexability for future updates (e.g. one script for each function).

>  In concstats_comp() and concstats_inequ(), normalize should be further documented and potentially checked.
>
> If normalize = NA, it gives a generic R warning. It seems that this should be checked along with the logical and length checks, as you are already checking those and providing a directly helpful error message.
> Further, normalize is summarily ignored if type = 'all'. This may be intended behavior, but should be documented or checked for. Otherwise, it confidently gives the wrong (or at least unexpected) answer when normalize = TRUE for the gini and entropy measures compared to when normalize = TRUE and type %in% c('gini', 'entropy') or type %in% c('hhi')

Response: Included checks/tests for normalized = NA. Functions like `concstats_all_* and functions with type = "all" have now a normalized = normalized argument

> Line 40 of R/comp.R references a function (concstats_all()) which does not exist. This should be revised to point to (presumably)concstats_all_comp().

> Line 39 of R/inequ.R references a function (concstats_all()) which does not exist. This should be revised to point to (presumably)concstats_all_inequ().

> Line 27 of R/mstruct.R references a function (concstats_all()) which does not exist. This should be revised to point to (presumably)concstats_all_mstruct().

Response: functions have been referenced as suggested, e.g. concstats_all_comp(), concstats_all_inequ(), concstats_all_mstruct

> Documentation for concstats_concstats() would benefit from the list of measures that will be computed. It indicates "a set of different and selected structural, inequality, and concentration measures" will be computed, but indicating which those are would be useful to users.

Response: All measures included in this set are now mentioned in the documentation/details

> x is documented as "@param x a numeric vector of length 1 with non-negative values." throughout R/comp.R, R/concstats.R, R/inequ.R, and R/mstruct.R. This seems incorrect and conflicts with the examples, which always show a numeric vector of positive length.

Response: This has been changed to "x A non-negative numeric vector."

> Input classes and lengths are consistently checked. One point of question: When checking x, what is the intended behavior if length(x) == 1? In some cases, this works. In others, such as for concstats_grs(), it fails with an uninformative error.

Response: This has been removed/changed to sum(x) == 1

> concstats_constats(), concstats_all_comp(), concstats_all_inequ(), and concstats_all_mstruct() each round the output to 2 decimal digits. This may be a reasonable choice, but it is not clear why 2 decimal digits is appropriate here, but not in the other functions. Should all outputs contain a rounding step or is this step unnecessary? One way to make this more consistent could be to make this optional, controlled by a new argument.

Response: All functions who return a table of the group functions now have an additional argument with default "digits = NULL" as an intermediary between precision and style.

> In each of concstats_all_comp(), concstats_all_inequ(), and concstats_all_mstruct(), there is code which runs via invisible(utils::capture.output(...)). It is not clear what the purpose of this is, as there does not appear to be any print statements in the functions in the .... If this is intended to capture some known issue with printing elsewhere, it would seem best to deal with this directly by modifying inputs or providing warnings or errors where appropriate.

Response: Advise on this will be apreciated. The intencion is to compute the individual functions but not return the results one by one, instead "collect" them and return them all in table. I haven't found a better way to code this yet.

> However, a concerning number of the tests in the package are inappropriate for testing the package. Hundreds of the tests are testing basic aspects of base R, not testing pieces of the package. Things like testing R's assignment operator, that NAs are NA, or that vectors sum properly. These tests should be removed.

Response: I do agree with respect to NAs, however, I kept tests to check if the vectors sums up to (near) 1, since, in my opinion, this is an essential characteristic of the vector. Probably you do agree that a relative market share of 90% or 130% make no sense. If I don't control for the input the functions will work, but the results will be erronous (due to copy-paste or typing errors). Furthermore, sometimes the user of the package is not the same person who prepared the raw data.

> In many calls to expect_error() across all test files it appears that the parentheses are frequently not matched properly, so that what appears to be an argument for regexp is instead passed to isTRUE() or to concstats_*(). That will always cause an error, so the test is again not testing aspects of the package, but instead that a base R function performs how it is intended. In many cases, this means that tests designated for particular standards with "@srrstats {G5.2, G5.2a, G5.2b, G5.8, G5.8a, G5.8b}" (e.g. in tests/testthat/test_comp.R l. 64-68) do not appear to test anything that would satisfy those standards for concstats. This is a pervasive problem in the tests that would need substantial revision.

Response: I have reviewed expect_error() (and other expectations) and the parentheses. Some of these were definitly wrong. Some have now this structure:
 act <- concstats_hhi(x)
 exp <- concstats_hhi(x4 / sum(x4))
 expect_equal(act, exp, tolerance = .Machine$double.eps^0.25)

Others:
expect_error(concstats_hhi(sum(x1), 1, tolerance = .Machine$double.eps^0.25))
to satisfy tests and "@srrstats" standards. In addition, in future updates each function will have its own file and tests (as recommended standard) with the intent to make this more readable and make testing easier.

> In data-raw/creditcoops.R, on line 3 n() should be dplyr::n(), otherwise the demo data cannot be reproduced (assuming no packages loaded beyond devtools::load_all()).

Response: has been changed to dplyr::n()

> Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

Response: Of course, both reviewers have been added to the DESCRIPTION file
