### 2023-03-15 #559 @christopherkenny
Hi @schneiderpy, thanks for this.  The changes look good to me. Thank you for your attention to the details here.

Re:
> Response: Advise on this will be apreciated. The intencion is to compute the individual functions but not return the results one by one, instead "collect" them and return them all in table. I haven't found a better way to code this yet.
It would seem to me that you can just remove the `invisible(utils::capture.output(...))` (and inner commas) as they will not return anything to the global environment. You control their return through the table later.

Some extra notes that *do not need to be addressed*, but may be helpful for future maintenance of the package:
- `@rdname` does not need to be specified when it is the name of a single function, as it defaults to the name of that function.
- Function reference can be categorized by type, [as documented here](https://pkgdown.r-lib.org/reference/build_reference.html#reference-index).
