### 2023-03-16 #559 @schneiderpy
Hi @christopherkenny and thank you for the advise. I have already tested it and it works fine.
I will consider the extra notes for future updates.
Again, thank you for reviewing the package and the valuable comments and advise.
