### 2023-03-17 #559 @schneiderpy
Thank you @Paula-Moraga for the good news!
I could not find any useful instruction on how to "transfer the repo to rOpenSci's "ropensci" GitHub organization".
Any advise would be apreciated to continue the process.
